Add Page support in Content directory
Instead of creating pages using Gatsby's File System Route API we now
create pages using gatsby-node file. This allows us to create pages
based on any criteria and any path.

This approach requires us to create templates for each kind of page type
we create. Because of that we created templates folder.

Also in gatsby-node file in onCreateNode hook, we added source field to
all Mdx nodes in order to find out which page we are creating currently
based on the name field in gatsby-source-filesystem plugin.